+++ packages/components/src/styles/component.styles.ts
@@ -1,8 +1,6 @@
-/* eslint-disable */
 import { css } from 'lit';
 
 export default css`
-	/* stylelint-disable */
   :host {
     box-sizing: border-box;
   }